biyesheji/requirements.txt
faker==19.13.0
mysql-connector-python==8.2.0
pandas==2.1.4
tqdm==4.66.1
python-dotenv==1.0.0 